Glue and paste Bokeh figures by HTML/JS output
This change adds the glue_bokeh function to insert the Bokeh figure
HTML and JS components into the cell output. The PasteBokeh handler then
collects the appropriate glued output and returns both nodes to be
included simultaneously.

The Bokeh figure components are returned as strings from the components
method, so they are converted to IPython HTML and Javascript objects
before being passed to glue() so that they have the correct mimetypes
for the display formatter. In addition, creating the Javascript object
allows us to add the BokehJS library dependency which will be injected
into the head of the HTML file.

This approach seems to be fairly robust, insofar as the figure is pasted
properly. It does not yet support being pasted as a MyST-NB Figure
though. Disadvantages of this approach, as I see them, are:

1) It requires a separate function to glue Bokeh figures, which isn't a
   super great UI, if it can be avoided
2) If a page has multiple pasted Bokeh figures, it isn't clear whether
   the BokehJS library will be injected into the header multiple times,
   and I'm not familiar enough with HTML semantics to know whether
   loading the same script multiple times is actually a problem.